src/os_support.c: fix sim_root_len type declaration
Type mismatch is detected by by '-flto -Werror=lto-type-mismatch':

    os_support.c:32:12: error: type of 'sim_root_len' does not match original declaration [-Werror=lto-type-mismatch]
       32 | extern int sim_root_len;
          |            ^
    load.c:104:8: note: type 'size_t' should match type 'int'
      104 | size_t sim_root_len = 0;
          |        ^

+++ src/os_support.c
@@ -29,7 +29,7 @@
 #include "os_support.h"
 
 extern char *sim_root;
-extern int sim_root_len;
+extern size_t sim_root_len;
 
 static SimTime simtime0;
